wk - change input stategy to match typerace to lessen typing burden
