--- src/sections/asteroidLogic/asteroidLogic.css
@@ -1,19 +1,0 @@
-.asteroid {
-    transform: translate(-50%, -50%);
-    width: fit-content;
-}
-
-.asteroid img {
-    position: absolute;
-
-    min-width: 100px;
-    width: 150%;
-
-    transform: translate(-50%, -50%);
-}
-
-.asteroid .asteroid-word {
-    transform: translate(-50%, -50%);
-
-    font-size: 1.5rem;
-}
